+++ src/Services/.BaseService/AppyNox.Services.Base.Application/Interfaces/Authentication/INoxTokenManager.cs
@@ -1,0 +1,8 @@
+﻿namespace AppyNox.Services.Base.Application.Interfaces.Authentication
+{
+    public interface INoxTokenManager
+    {
+        string GetUserInfoByToken(string token);
+        bool VerifyToken(string token);
+    }
+}