+++ src/Services/AuthenticationService/AppyNox.Services.Authentication.Application/DTOs/ApplicationRoleDTOs/Models/ApplicationRoleCreateDto.cs
@@ -1,0 +1,16 @@
+﻿using AppyNox.Services.Base.Application.Dtos;
+
+namespace AppyNox.Services.Authentication.Application.DTOs.ApplicationRoleDTOs.Models
+{
+    /// <summary>
+    /// Data transfer object for creating a new identity role.
+    /// </summary>
+    public class ApplicationRoleCreateDto : BaseDto
+    {
+        #region [ Properties ]
+
+        public string Name { get; set; } = string.Empty;
+
+        #endregion
+    }
+}