+++ src/Services/AuthenticationService/AppyNox.Services.Authentication.Application/DTOs/RefreshTokenDtos/Models/RefreshTokenDto.cs
@@ -1,0 +1,16 @@
+﻿namespace AppyNox.Services.Authentication.Application.DTOs.RefreshTokenDtos.Models
+{
+    /// <summary>
+    /// Data transfer object representing a refresh token.
+    /// </summary>
+    public class RefreshTokenDto
+    {
+        #region [ Properties ]
+
+        public string Token { get; set; } = string.Empty;
+
+        public string RefreshToken { get; set; } = string.Empty;
+
+        #endregion
+    }
+}