+++ src/Services/AuthenticationService/AppyNox.Services.Authentication.Domain/Entities/ApplicationUser.cs
@@ -2,9 +2,15 @@
 
 namespace AppyNox.Services.Authentication.Domain.Entities
 {
-    public class ApplicationUser : IdentityUser
+    public class ApplicationUser : IdentityUser<Guid>
     {
         #region [ Properties ]
+
+        public bool IsAdmin { get; set; }
+
+        #endregion
+
+        #region [ Relations ]
 
         public Guid CompanyId { get; set; }